+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The resulting web page also makes it easy for users to search the table for inte
 * 'Ethernet' filter features are now mutually exclusives
 * Fixes Wlan 60Ghz Title typo (was 600Ghz)
 * Displays Filtered / Total count in the header
+* Links Header Title to TOH's base URL
 
 -->